feat(utxo-lib): move some testutils to `src/`
These functions are useful in tests of other packages.

Issue: BG-63135
