+++ modules/utxo-lib/src/testutil/index.ts
@@ -1,0 +1,2 @@
+export * from './keys';
+export * from './mock';
