+++ modules/utxo-lib/test/integration_local_rpc/generate/fixtures.ts
@@ -3,9 +3,9 @@ import * as fs from 'fs-extra';
 import * as path from 'path';
 
 import { networks, Network, getNetworkName, getMainnet } from '../../../src';
+import { getKeyTriple } from '../../../src/testutil';
 import { RpcClient } from './RpcClient';
 import { RpcTransaction } from './RpcTypes';
-import { getKeyTriple } from '../../testutil';
 import { getDefaultTransactionVersion } from '../../../src/bitgo';
 
 export type Protocol = {
