+++ modules/utxo-lib/test/testutil/index.ts
@@ -1,2 +1,1 @@
-export * from './keys';
 export * from './tnumber';
